chore(deps): bump rusqlite 0.34 → 0.40 (+ libsqlite3-sys 0.32 → 0.38)
Unblocks the oxiline integration: rusqlite 0.40 satisfies both oxios's
memory/sqlite-vec stack and oxiline's rusqlite_migration 2.x (which needs
rusqlite >= 0.37). sqlite-vec 0.1.9 has NO runtime rusqlite dependency
(FFI only; rusqlite is [dev-dependencies] there), so the only constraint
was oxios's own pin.

Adapts to rusqlite 0.40 removing the usize/u64 ToSql/FromSql impls
(SQLite integers are i64; usize/u64 can exceed it): cast to i64 at the
8 affected bind/read sites (oxios-memory dream/bm25/vector/store,
oxios-kernel task store + agent_log_db).

Verified: cargo check/clippy (-D warnings) clean (default + memo);
cargo test --workspace fully green (memory 210+219, kernel 825, etc.).

Also bumps the oximemo-core path pin 0.6 → 0.7 (local crate moved on).

+++ crates/oxios-memory/src/memory/sqlite/search/bm25.rs
@@ -88,7 +88,7 @@ pub fn search_bm25(db: &MemoryDatabase, query: &str, limit: usize) -> Result<Vec
             return Ok(Vec::new());
         }
     };
-    let rows = match stmt.query_map(rusqlite::params![fts_query, limit], |row| {
+    let rows = match stmt.query_map(rusqlite::params![fts_query, limit as i64], |row| {
         Ok(Bm25Hit {
             rowid: row.get(0)?,
             id: row.get(1)?,
